refactor engine into three modules
It was getting to messy. Also, I've dropped some bad error handling and
switched to panicing. Unfortunately, threading through the error was
really nasty.
